# [FIX] base_address_extended: add complete address to contact