@xmo-odoo 2021-04-08
> edit: but maybe the performance improvement just comes from ` and 'street' not in vals:`

According to the commit message both are involved:

> First remove syncing extended fields, that seems completely unnecessary since the street gets sync'd and will get split through the normal process (updating the sub-fields), **by also syncing the split fields we're redundantly calling _set_steet and rewriting the address we just sync'd**.

Meaning this triggered extra `_set_street` and thus `write` calls.

Sadly I apparently did not take note of the performance change for this specific change, and I can't say I remember this PR at all, but this change needs to be benchmarked and profiled to ensure it doesn't regress the original issue (#31549, the fix PR is at #31876).