@nboulif 2021-05-25
ping @xmo-odoo