@nboulif 2021-05-31
Hello @xmo-odoo , do you think this PR is ready to be merged? Thanks (again) in advance!