@xmo-odoo 2021-06-01
@nboulif sure? well I've got no objection anyway.