@nboulif 2021-06-01
@xmo-odoo I think so :slightly_smiling_face: 
@nle-odoo Hello, can I ask you to do a last check please? Thanks in advance!