@nboulif 2021-06-02
@nle-odoo Thanks!
> I guess: `if partner.street!= street_value` is to prevent just the write or that the write cause another code to be executed?

Indeed, It cause another code to be executed ( the recompute of extended street_fields ) :slightly_smiling_face: