@naglis 2021-09-08
Hi, is there any chance this will be forward ported to Odoo 14.0?